+++ test/run_all_versions.sh
@@ -108,6 +108,7 @@ SUITES=(
 	native_agg
 	native_agg_addcolumn
 	native_agg_deletes
+	native_batch_fold_projection
 	native_backend_crash
 	native_bloom
 	native_cancel